requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
strawberry-graphql==0.205.0
pydantic==2.3.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
rich==13.5.2
psycopg2-binary==2.9.7

